requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
python-multipart
aiofiles
google-genai

